Issue #604: updated unit tests
